+++ changelog.md
@@ -13,6 +13,8 @@ for a version 1.0 release in early 2
   - Extension: Strikethrough
 - [Issue 805](https://github.com/jackdewinter/pymarkdown/issues/805)
   - Extension: Task List Items
+- [Issue 825](https://github.com/jackdewinter/pymarkdown/issues/825)
+  - Rule MD048 - Added fix options
 
 ### Changed